Use prettier to format code
Before this change there were a lot of warnings which made it difficult
to spot errors in log output. Moreover, the support for auto formatting
was outdated. By using prettier, we were able to resolve most of the
warnings.

By using prettier it's less common for lint errors to take place.
Becuase of that, I moved lint checks from the pre-commit hook into the
CI.