Fix componentDidUpdate logic
During an earlier refactor which replaced a deprecated lifecycle
method with componentDidUpdate, the logic within this function broke.

The logic of the lifecycle method should check if the unit has changed
or been selected--if so it should center the map. The previous version
made this logical comparison against the previous unit when it should
have done so against the current unit.