Double encode unit name
The current version of history decodes paths before passing
them on to the history API. By double encoding, we ensure that
an encoded version of the path is saved into state.

This is fixed in history v5, but it requires react-router v6,
which would be a more major change, and v6 was not yet released
at the time of this change.